start to make the menu for dragon city
- Some change to lib due to bugs
- created menu for dragon city

+++ Xna2DLib/Animation/DoubleAnimation.cs
@@ -14,7 +14,7 @@ namespace tranduytrung.Xna.Animation
         public override void Initialize()
         {
             if (From == null)
-                ActualFrom = (double)GetAccessor.Invoke();
+                ActualFrom = GetAccessor.Invoke();
             else
                 ActualFrom = From.Value;
         }
